refactor: move flat types next to their entity modules
Flat types are a general serialization concept, not tied to caching.
Move each flat type and conversion util to its entity's module:
- FlatWorkspace → core-modules/workspace/types/
- FlatApiKey → core-modules/api-key/types/
- FlatUserWorkspace → core-modules/user-workspace/types/
- FlatAuthContextUser → core-modules/auth/types/

Made-with: Cursor
