+++ packages/twenty-server/@types/express.d.ts
@@ -1,9 +1,9 @@
 import { type APP_LOCALES } from 'twenty-shared/translations';
 
-import { type FlatApiKey } from 'src/engine/core-entity-cache/types/flat-api-key.type';
-import { type FlatAuthContextUser } from 'src/engine/core-entity-cache/types/flat-auth-context-user.type';
-import { type FlatUserWorkspace } from 'src/engine/core-entity-cache/types/flat-user-workspace.type';
-import { type FlatWorkspace } from 'src/engine/core-entity-cache/types/flat-workspace.type';
+import { type FlatApiKey } from 'src/engine/core-modules/api-key/types/flat-api-key.type';
+import { type FlatAuthContextUser } from 'src/engine/core-modules/auth/types/flat-auth-context-user.type';
+import { type FlatUserWorkspace } from 'src/engine/core-modules/user-workspace/types/flat-user-workspace.type';
+import { type FlatWorkspace } from 'src/engine/core-modules/workspace/types/flat-workspace.type';
 import { type ApplicationEntity } from 'src/engine/core-modules/application/application.entity';
 import { type RawAuthContext } from 'src/engine/core-modules/auth/types/auth-context.type';
 import { type AuthProviderEnum } from 'src/engine/core-modules/workspace/types/workspace.type';
