+++ packages/twenty-server/src/engine/api/graphql/graphql-config/interfaces/graphql-context.interface.ts
@@ -1,7 +1,7 @@
 import { type YogaDriverServerContext } from '@graphql-yoga/nestjs';
 
-import { type FlatAuthContextUser } from 'src/engine/core-entity-cache/types/flat-auth-context-user.type';
-import { type FlatWorkspace } from 'src/engine/core-entity-cache/types/flat-workspace.type';
+import { type FlatAuthContextUser } from 'src/engine/core-modules/auth/types/flat-auth-context-user.type';
+import { type FlatWorkspace } from 'src/engine/core-modules/workspace/types/flat-workspace.type';
 
 export interface GraphQLContext extends YogaDriverServerContext<'express'> {
   user?: FlatAuthContextUser;
