+++ packages/twenty-server/src/engine/api/mcp/controllers/mcp-core.controller.ts
@@ -18,8 +18,8 @@ import { JsonRpc } from 'src/engine/api/mcp/dtos/json-rpc';
 import { McpAuthGuard } from 'src/engine/api/mcp/guards/mcp-auth.guard';
 import { McpProtocolService } from 'src/engine/api/mcp/services/mcp-protocol.service';
 import { RestApiExceptionFilter } from 'src/engine/api/rest/rest-api-exception.filter';
-import { FlatApiKey } from 'src/engine/core-entity-cache/types/flat-api-key.type';
-import { FlatWorkspace } from 'src/engine/core-entity-cache/types/flat-workspace.type';
+import { FlatApiKey } from 'src/engine/core-modules/api-key/types/flat-api-key.type';
+import { FlatWorkspace } from 'src/engine/core-modules/workspace/types/flat-workspace.type';
 import { UserEntity } from 'src/engine/core-modules/user/user.entity';
 import { AuthApiKey } from 'src/engine/decorators/auth/auth-api-key.decorator';
 import { AuthUserWorkspaceId } from 'src/engine/decorators/auth/auth-user-workspace-id.decorator';
