+++ packages/twenty-server/src/engine/core-modules/api-key/services/workspace-api-key-map-cache.service.ts
@@ -5,8 +5,8 @@ import { Repository } from 'typeorm';
 
 import { WorkspaceCacheProvider } from 'src/engine/workspace-cache/interfaces/workspace-cache-provider.service';
 
-import { type FlatApiKey } from 'src/engine/core-entity-cache/types/flat-api-key.type';
-import { fromApiKeyEntityToFlat } from 'src/engine/core-entity-cache/utils/from-api-key-entity-to-flat.util';
+import { type FlatApiKey } from 'src/engine/core-modules/api-key/types/flat-api-key.type';
+import { fromApiKeyEntityToFlat } from 'src/engine/core-modules/api-key/utils/from-api-key-entity-to-flat.util';
 import { ApiKeyEntity } from 'src/engine/core-modules/api-key/api-key.entity';
 import { WorkspaceCache } from 'src/engine/workspace-cache/decorators/workspace-cache.decorator';
