+++ packages/twenty-server/src/engine/core-modules/api-key/utils/from-api-key-entity-to-flat.util.ts
@@ -1,5 +1,5 @@
 import { type ApiKeyEntity } from 'src/engine/core-modules/api-key/api-key.entity';
-import { type FlatApiKey } from 'src/engine/core-entity-cache/types/flat-api-key.type';
+import { type FlatApiKey } from 'src/engine/core-modules/api-key/types/flat-api-key.type';
 
 export const fromApiKeyEntityToFlat = (entity: ApiKeyEntity): FlatApiKey => ({
   id: entity.id,
