+++ packages/twenty-server/src/engine/core-modules/auth/token/services/workspace-agnostic-token.service.ts
@@ -5,7 +5,7 @@ import { addMilliseconds } from 'date-fns';
 import ms from 'ms';
 import { Repository } from 'typeorm';
 
-import { fromAuthContextUserToFlat } from 'src/engine/core-entity-cache/utils/from-auth-context-user-to-flat.util';
+import { fromAuthContextUserToFlat } from 'src/engine/core-modules/auth/utils/from-auth-context-user-to-flat.util';
 import {
   AuthException,
   AuthExceptionCode,
