+++ packages/twenty-server/src/engine/core-modules/auth/types/flat-auth-context-user.type.ts
