+++ packages/twenty-server/src/engine/core-modules/user-workspace/types/flat-user-workspace.type.ts
