+++ packages/twenty-server/src/engine/core-modules/workspace/services/workspace-entity-cache-provider.service.ts
@@ -5,8 +5,8 @@ import { Repository } from 'typeorm';
 
 import { CoreEntityCache } from 'src/engine/core-entity-cache/decorators/core-entity-cache.decorator';
 import { CoreEntityCacheProvider } from 'src/engine/core-entity-cache/interfaces/core-entity-cache-provider.service';
-import { type FlatWorkspace } from 'src/engine/core-entity-cache/types/flat-workspace.type';
-import { fromWorkspaceEntityToFlat } from 'src/engine/core-entity-cache/utils/from-workspace-entity-to-flat.util';
+import { type FlatWorkspace } from 'src/engine/core-modules/workspace/types/flat-workspace.type';
+import { fromWorkspaceEntityToFlat } from 'src/engine/core-modules/workspace/utils/from-workspace-entity-to-flat.util';
 import { WorkspaceEntity } from 'src/engine/core-modules/workspace/workspace.entity';
 
 @Injectable()
