+++ packages/twenty-server/src/engine/core-modules/workspace/types/flat-workspace.type.ts
