+++ packages/twenty-server/src/engine/twenty-orm/utils/build-system-auth-context.util.ts
@@ -1,6 +1,6 @@
 import { type SystemWorkspaceAuthContext } from 'src/engine/core-modules/auth/types/workspace-auth-context.type';
 import { buildSystemAuthContext as buildSystemAuthContextUtil } from 'src/engine/core-modules/auth/utils/build-system-auth-context.util';
-import { type FlatWorkspace } from 'src/engine/core-entity-cache/types/flat-workspace.type';
+import { type FlatWorkspace } from 'src/engine/core-modules/workspace/types/flat-workspace.type';
 
 // Builds a minimal WorkspaceAuthContext for system operations (jobs, commands, crons)
 // that don't have a user context but need to operate on a workspace
